requirements.txt
# MATLABバージョン: 2020b
# system identification toolbox が必要
#
#
# 以下はpython環境 
# 
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=20.1.0=py38h2bbff1b_1
astroid=2.4.2=py38_0
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.4=py_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
ca-certificates=2021.4.13=haa95532_1
certifi=2020.12.5=py38haa95532_0
cffi=1.14.5=py38hcd4344a_0
colorama=0.4.3=py_0
cudatoolkit=10.2.89=h74a9793_1
cycler=0.10.0=py38_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
flake8=3.8.3=py_0
freetype=2.10.2=hd328e21_0
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
importlib-metadata=1.7.0=py38_0
importlib_metadata=1.7.0=hd3eb1b0_0
intel-openmp=2020.1=216
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=4.3.21=py38_0
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.2.0=py38h74a9793_0
lazy-object-proxy=1.4.3=py38he774522_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=h62dcd97_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
matlabengineforpython=R2020b=pypi_0
matplotlib=3.2.2=0
matplotlib-base=3.2.2=py38h64f37c6_0
mccabe=0.6.1=py38_1
mistune=0.8.4=py38he774522_1000
mkl=2020.1=216
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.1.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
mypy=0.782=py_0
mypy_extensions=0.4.3=py38_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
ninja=1.10.0=py38h7ef1ec2_0
notebook=6.3.0=py38haa95532_0
numpy=1.19.1=py38h5510c5b_0
numpy-base=1.19.1=py38ha3acd2a_0
olefile=0.46=py_0
openssl=1.1.1k=h2bbff1b_0
packaging=20.9=pyhd3eb1b0_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=7.2.0=py38hcc1f983_0
pip=20.2.2=py38_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.7.0=py38he774522_0
pycodestyle=2.6.0=py_0
pycparser=2.20=py_2
pyflakes=2.2.0=py_0
pygments=2.8.1=pyhd3eb1b0_0
pylint=2.5.3=py38_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38_1
pyrsistent=0.17.3=py38he774522_0
python=3.8.5=he1778fa_0
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.8_cuda102_cudnn7_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyzmq=20.0.0=py38hd77b12b_1
qt=5.9.7=vc14h73c81de_0
scipy=1.5.0=py38h9439919_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=49.6.0=py38_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py_0
sqlite=3.32.3=h2a8f88b_0
terminado=0.9.4=py38haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
tk=8.6.10=he774522_0
toml=0.10.1=py_0
torchvision=0.7.0=py38_cu102
tornado=6.0.4=py38he774522_1
traitlets=5.0.5=pyhd3eb1b0_0
typed-ast=1.4.1=py38he774522_0
typing_extensions=3.7.4.2=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.34.2=py38_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
wrapt=1.11.2=py38he774522_0
xz=5.2.5=h62dcd97_0
zeromq=4.3.3=ha925a31_3
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
